Fix FAISS float32 dtype, replace examples with RecA and CoxA
- Fix: cast GPU embeddings to float32 (FAISS requires it)
- Replace EGFR/P53 examples with RecA and CoxA (E. coli)
- Keep Insulin example

+++ modal_app.py
@@ -274,7 +274,7 @@ def ui():
         result = embedder.embed.remote(sequences)
         if progress:
             progress(0.9, desc="Embeddings received from GPU!")
-        return np.array(result)
+        return np.array(result, dtype=np.float32)
 
     gi.run_embed_protein_vec = gpu_embed
